use loader meachanism for views
