+++ src/frontend/src/Loading.tsx
@@ -1,6 +1,7 @@
 import { Center, Loader, Stack } from '@mantine/core';
+import { Suspense, lazy } from 'react';
 
-export default function Loading() {
+export function Loading() {
   return (
     <Stack>
       <Center>
@@ -9,3 +10,10 @@ export default function Loading() {
     </Stack>
   );
 }
+
+export const Loadable = (Component: any) => (props: JSX.IntrinsicAttributes) =>
+  (
+    <Suspense fallback={<Loading />}>
+      <Component {...props} />
+    </Suspense>
+  );
