+++ src/frontend/src/context/BaseContext.tsx
@@ -1,0 +1,10 @@
+import { LanguageContext } from './LanguageContext';
+import { ThemeContext } from './ThemeContext';
+
+export const BaseContext = ({ children }: { children: any }) => {
+  return (
+    <LanguageContext>
+      <ThemeContext>{children}</ThemeContext>
+    </LanguageContext>
+  );
+};
